# backend/static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DiabeCare API",
    "description": "DiabeCare \u2014 Diabetic Patient Medication Follow-Up System API",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://your-production-domain.com/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:5000/api",
      "description": "Local dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "admin"
                  },
                  "password": {
                    "type": "string",
                    "example": "admin123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful \u2014 returns JWT token and user object"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get current authenticated user",
        "responses": {
          "200": {
            "description": "Current user object"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a new user (admin only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "email",
                  "password",
                  "role",
                  "full_name"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "jdoe"
                  },
                  "email": {
                    "type": "string",
                    "example": "jdoe@email.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "pass1234"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "doctor",
                      "patient"
                    ]
                  },
                  "full_name": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "phone": {
                    "type": "string",
                    "example": "+254700000000"
                  },
                  "age": {
                    "type": "integer",
                    "description": "Required when role=patient"
                  },
                  "gender": {
                    "type": "string",
                    "enum": [
                      "Male",
                      "Female"
                    ]
                  },
                  "diabetes_type": {
                    "type": "string",
                    "enum": [
                      "Type 1",
                      "Type 2",
                      "Gestational"
                    ]
                  },
                  "blood_sugar": {
                    "type": "number"
                  },
                  "hba1c": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Missing required fields"
          },
          "403": {
            "description": "Admin access required"
          },
          "409": {
            "description": "Username or email already exists"
          }
        }
      }
    },
    "/auth/users": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "List all users (admin only)",
        "responses": {
          "200": {
            "description": "Array of user objects"
          },
          "403": {
            "description": "Admin access required"
          }
        }
      }
    },
    "/auth/users/{user_id}": {
      "put": {
        "tags": [
          "Auth"
        ],
        "summary": "Update a user (admin only)",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "full_name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "doctor",
                      "patient"
                    ]
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated"
          },
          "403": {
            "description": "Admin access required"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Username or email conflict"
          }
        }
      },
      "delete": {
        "tags": [
          "Auth"
        ],
        "summary": "Delete a user (admin only)",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted"
          },
          "400": {
            "description": "Cannot delete self or last admin"
          },
          "403": {
            "description": "Admin access required"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/auth/change-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Change own password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "current_password",
                  "new_password"
                ],
                "properties": {
                  "current_password": {
                    "type": "string"
                  },
                  "new_password": {
                    "type": "string",
                    "description": "Minimum 8 characters"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed"
          },
          "400": {
            "description": "Missing fields or too short"
          },
          "401": {
            "description": "Current password incorrect"
          }
        }
      }
    },
    "/auth/profile": {
      "put": {
        "tags": [
          "Auth"
        ],
        "summary": "Update own profile (full_name, email, phone)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "full_name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated \u2014 returns updated user object"
          }
        }
      }
    },
    "/auth/avatar": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Upload or remove profile avatar (base64 data-URI)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar_url": {
                    "type": "string",
                    "description": "Base64 data-URI (e.g. data:image/png;base64,...). Pass empty string to remove.",
                    "example": "data:image/jpeg;base64,/9j/4AAQ..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Avatar updated \u2014 returns avatar_url"
          },
          "400": {
            "description": "Invalid format"
          },
          "413": {
            "description": "Image too large (max ~2 MB)"
          }
        }
      }
    },
    "/auth/activity-log": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get audit/activity log (admin only) \u2014 last 200 entries",
        "responses": {
          "200": {
            "description": "Array of audit log entries"
          },
          "403": {
            "description": "Admin access required"
          }
        }
      }
    },
...
  },
...
}